+++ parseq_XAS/XAS_fits.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 u"""
-Data fits
----------
+No GUI: data fits
+-----------------
 
 Linear Combination Fit of µ(E)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -32,21 +32,21 @@ class LCF_mu(fits.lcf.LCF):
     The Linear Combination Fit (LCF) fits a weighted sum of reference spectra
     to the spectrum of interest. Here, the array attribute :param:`flat` (the
     version of µ that has a horizontal post-edge) of the involved data objects
-    are used as ordinates and the attributes :param:`e` are used as abscissas.
-    The fit curve resides in a new array :param:`fitLCF` as an attribute of the
+    are used as ordinates and the attribute :param:`e` is used as abscissa. The
+    fit curve resides in a new array :param:`fitLCF` as an attribute of the
     data container.
 
-    The fit creates an attrubute :param:`lcf_params` of the data container that
-    is a list of dictionaries, one dictionary per reference spectrum, with the
+    The fit creates an attribute :param:`lcf_params` of the data container that
+    is a list of dictionaries, a dictionary per reference spectrum, with the
     following items:
-
-        *name*: str, the alias of the reference spectrum,
-        *use*: bool, switches the reference on/off,
-        *w*: float, the weight of the reference,
-        *wBounds*: 3-list, [*min*, *max*, *Δ*], Δ is only used by the GUI,
-        If :param:`xVary` is True (see below), these two entries are needed:
-        *dx*: float, the energy shift of the reference,
-        *dxBounds*: 3-list, [*min*, *max*, *Δ*] for the energy shift.
+       | *name*: str, the alias of the reference spectrum,
+       | *use*: bool, switches the reference on/off,
+       | *w*: float, the weight of the reference,
+       | *wBounds*: 3-list, [*min*, *max*, *Δ*], Δ is only used by the GUI,
+       |
+       | If :param:`xVary` is True (see below), these two entries are needed:
+       | *dx*: float, the energy shift of the reference,
+       | *dxBounds*: 3-list, [*min*, *max*, *Δ*] for the energy shift.
 
     The reference dictionary may have *wtie* and *dxtie* for tie expressions
     for the weight *w* and the energy shift *dx*. Tie expression is a string
@@ -63,13 +63,13 @@ class LCF_mu(fits.lcf.LCF):
     are allowed to vary their energy axis. This feature may be needed if the
     energy calibration of the involved spectra is not trustful.
 
-    The actual curve fitting is done by `scipy.optimize.curve_fit()`. The
-    attribute :param:`lcf_result` of the data container will get fit info
-    returned by `curve_fit()`.
+    The actual curve fitting is done by ``scipy.optimize.curve_fit()``. The
+    attribute :param:`lcf_result` of the data container will get the fit info
+    returned by ``curve_fit()``.
     """
 
     name = "LCF"
-    ref = "fits.html#linear-combination-fit-of-e"
+    ref = "nogui.html#linear-combination-fit-of-e"
     tooltip = "Linear Combination Fit. Use a 'flat' normalized mu view."
     xVary = True
     dataAttrs = dict(x='e', y='flat', fit='fitLCF')
@@ -89,23 +89,24 @@ class FunctionFit_mu(fits.functionfit.FunctionFit):
     resides in a new array :param:`fitFunc` as an attribute of the data
     container.
 
-    The fit creates a few attrubutes of the data container:
+    The fit creates a few attributes of the data container:
 
         * :param:`ffit_formula`: str, fit formula,
         * :param:`ffit_params`: dictionary of fitting variables as keys,
         * :param:`ffit_xRange`: 2-list [*min*, *max*] of energy limits.
 
     The dictionary :param:`ffit_params` is a dictionary of dictionaries of the
-    following keys: *value* (float), *lim* (2-list of *min* and *max*), *tie*
-    (str, a tie expression) and *error* (float, the fitting error).
+    following keys: *value* (float, the fit variable value), *lim* (2-list of
+    *min* and *max*), *tie* (str, a tie expression) and *error* (float, the
+    fitting error).
 
-    The actual curve fitting is done by `scipy.optimize.curve_fit()`. The
-    attribute :param:`ffit_result` of the data container will get fit info
-    returned by `curve_fit()`.
+    The actual curve fitting is done by ``scipy.optimize.curve_fit()``. The
+    attribute :param:`ffit_result` of the data container will get the fit info
+    returned by ``curve_fit()``.
     """
 
     name = "function fit"
-    ref = "fits.html#function-fit-of-e"
+    ref = "nogui.html#function-fit-of-e"
     tooltip = "Function fit with arbitrary formula. Use a "\
         "'flat' normalized mu view."
     dataAttrs = dict(x='e', y='flat', fit='fitFunc')
@@ -117,14 +118,14 @@ class FunctionFit_mu(fits.functionfit.FunctionFit):
 
 class EXAFSFit(fits.exafsfit.EXAFSFit):
     """
-    The EXAFS Fit fits a sum of EXAFS shells to to data. The array attribute
+    The EXAFS Fit fits a sum of EXAFS shells to data. The array attribute
     :param:`bft` of the fitted data object is used as ordinate and the
     attribute :param:`bftk` is used as abscissa. The fit curve resides in a new
     array :param:`bftfit` as an attribute of the data container. Optionally,
     the fit can be done in the Fourier space; then the fit works with
     :param:`ft` vs :param:`r` and the fit curve is :param:`ftfit`.
 
-    The fit creates a few attrubutes of the data container:
+    The fit creates a few attributes of the data container:
 
         * :param:`exafsfit_params`: list of shell dictionaries,
         * :param:`exafsfit_aux`: list of shell specifications -- a path to
@@ -142,18 +143,16 @@ class EXAFSFit(fits.exafsfit.EXAFSFit):
     ('tie': expression) entry.
 
     The list :param:`exafsfit_params` has an extra shell dictionary that
-    contains a 's0' fitting variable (multi-electron damping factor) and
+    contains an 's0' fitting variable (multi-electron damping factor) and
     optional metavariables.
 
-    The actual curve fitting is done by `scipy.optimize.curve_fit()`. The
-    attribute :param:`exafsfit_result` of the data container will get fit info
-    returned by `curve_fit()`.
-
-    Two kinds of fitting errors are offered. TODO
+    The actual curve fitting is done by ``scipy.optimize.curve_fit()``. The
+    attribute :param:`exafsfit_result` of the data container will get the fit
+    info returned by ``curve_fit()``.
     """
 
     name = "EXAFS fit"
-    ref = "fits.html#exafs-fit-of-k-and-r"
+    ref = "nogui.html#exafs-fit-of-k-and-r"
     tooltip = "EXAFS fit in filtered k- and/or in r-space"
     dataAttrs = dict(x='bftk', y='bft', fit='bftfit',
                      x2='r', y2='ft', fit2='ftfit')